requirements.txt
# Core Dependencies
streamlit>=1.28.0
crewai>=0.1.0
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Data Processing
numpy>=1.24.0

# Web Scraping Support
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Optional but recommended
plotly>=5.15.0
altair>=5.0.0

# Optional MCP Integration (commented because not pip-installable by default)
# mcp>=0.1.0
# crewai-tools>=0.1.0

# NLP Support (Spacy)
spacy>=3.7.0
